refactor: split misc.rs into log, notify, filelist, host modules
